Fix some really annoying sign-extension bugs.
The latter (LoadHalf) sign-extension bug caused severe clipping
issues. Applying this commit results in a much better picture.

+++ Memory.c
@@ -219,7 +219,7 @@ LoadHalf(const struct RSPMemoryData *memoryData, uint8_t *dmem) {
 
   /* Load and sign extend. */
   memcpy(&half, dmem + offset, sizeof(half));
-  *target = (int32_t) ByteOrderSwap16(half);
+  *target = (int32_t) ((int16_t) ByteOrderSwap16(half));
 }
 
 /* ============================================================================
